DoubleLinked Blockchain first part
This is the first part of the doublelinked blockchain.
Now ever block has two hashes, one with and one without transactions.
Furthermore, the blocks have to previous hashes as well, one with and
one without transactions.

Furutre Work:
- Enable searching whenever somewhere is searched something with
block.prevHash, becuaes it is also possible that now the
prevHashWithoutTx is valid.
- Fix tests