Next step in double linked blockchain
Now all other blokcs than the last three and the genesis block get moved
to the 'closedblockwithouttx' bucket.
The blocks now have a variable 'Aggregated'. This shows (later) if all
transactions are aggregated. If so the blcok is also moved to the
'closedblockwithouttx' bucket.
Furthermore some automatic cleanup from GOLAND.

Future Work:
- Ensure That blocks are found when requested from teh network.
- comment code
- Leader / aggregator distinction (if needed)