First step: aggregation of FundsTx in new block
In this commit the Aggregation Type AggTxSender get inserted. These
transactions are currently used for fundsTx Aggregaton in the current
block. Later the funds Tx will not be written into the
b.FundsTxDataSlice directly, but in an intermediate storage. From there
they will be aggregated and tehn inserted. This allows the possibility
to aggregate two or more transactions in one AggTxSender.

Future Work:
- The actual Aggregation
- comment code
- Rollbacks
- Exchange of AggTxSender between miners