Merging Fixed --> Multiple miner startup works
This commit brings the fix for starting up multiple miners with empty
blocks in the chain. Currently all blocks without transactions are moved
to the bucket without transactions included when they are out of the 3
blocks blocking period.
The request for blocks does now have both hashes included.
For the startup the all blocks are sorted according to the block.Height
and when the node is not bootstrapping, it does invert the blocks array.

Future Work:
- Conflicting Block Hash 1 & 2 need do be changed --> marked as TODO
- Historic transaction aggregation despite only empty blocks are moved.
- Check Rollback scenario once it occurs.